+++ Source/UZKArchive.m
@@ -462,7 +462,20 @@ NS_DESIGNATED_INITIALIZER
 
 - (BOOL)extractFilesTo:(NSString *)destinationDirectory
              overwrite:(BOOL)overwrite
-              progress:(void (^)(UZKFileInfo *currentFile, CGFloat percentArchiveDecompressed))progress
+                 error:(NSError * __autoreleasing*)error
+{
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
+    return [self extractFilesTo:destinationDirectory
+                      overwrite:overwrite
+                       progress:nil
+                          error:error];
+#pragma clang diagnostic pop
+}
+
+- (BOOL)extractFilesTo:(NSString *)destinationDirectory
+             overwrite:(BOOL)overwrite
+              progress:(void (^)(UZKFileInfo *currentFile, CGFloat percentArchiveDecompressed))progressBlock
                  error:(NSError * __autoreleasing*)error
 {
     UZKCreateActivity("Extracting Files to Directory");
@@ -483,8 +496,12 @@ NS_DESIGNATED_INITIALIZER
     NSFileManager *fm = [[NSFileManager alloc] init];
 
     NSNumber *totalSize = [fileInfo valueForKeyPath:@"@sum.uncompressedSize"];
-    UZKLogDebug("totalSize: %ld", totalSize.longValue);
+    UZKLogDebug("totalSize: %lld", totalSize.longLongValue);
     __block long long bytesDecompressed = 0;
+    __block NSInteger filesExtracted = 0;
+
+    NSProgress *progress = [self beginProgressOperation:totalSize.longLongValue];
+    progress.kind = NSProgressKindFile;
 
     __weak UZKArchive *welf = self;
     NSError *extractError = nil;
@@ -498,9 +515,18 @@ NS_DESIGNATED_INITIALIZER
             for (UZKFileInfo *info in fileInfo) {
                 UZKLogDebug("Extracting %{public}@ to disk", info.filename);
 
+                if (progress.isCancelled) {
+                    NSString *detail = [NSString localizedStringWithFormat:NSLocalizedStringFromTableInBundle(@"Error locating file '%@' in archive", @"UnzipKit", _resources, @"Detailed error string"),
+                                        info.filename];
+                    UZKLogError("Halted file extraction due to user cancellation: %{public}@", detail);
+                    [welf assignError:&strongError code:UZKErrorCodeUserCancelled
+                               detail:detail];
+                    return;
+                }
+                
                 @autoreleasepool {
-                    if (progress) {
-                        progress(info, bytesDecompressed / totalSize.doubleValue);
+                    if (progressBlock) {
+                        progressBlock(info, bytesDecompressed / totalSize.doubleValue);
                     }
                     
                     if (![self locateFileInZip:info.filename error:&strongError]) {
@@ -552,6 +578,10 @@ NS_DESIGNATED_INITIALIZER
                     
                     NSURL *deflatedDirectoryURL = [NSURL fileURLWithPath:destinationDirectory];
                     NSURL *deflatedFileURL = [deflatedDirectoryURL URLByAppendingPathComponent:info.filename];
+                    [progress setUserInfoObject:deflatedFileURL
+                                         forKey:NSProgressFileURLKey];
+                    [progress setUserInfoObject:info
+                                         forKey:UZKProgressInfoKeyFileInfoExtracting];
                     NSString *path = deflatedFileURL.path;
                     
                     UZKLogDebug("Creating empty file at path %{public}@", path);
@@ -590,8 +620,8 @@ NS_DESIGNATED_INITIALIZER
                                         UZKLogDebug("Writing data chunk of size %lu (%lld total so far)", (unsigned long)dataChunk.length, bytesDecompressed);
                                         bytesDecompressed += dataChunk.length;
                                         [deflatedFileHandle writeData:dataChunk];
-                                        if (progress) {
-                                            progress(info, bytesDecompressed / totalSize.doubleValue);
+                                        if (progressBlock) {
+                                            progressBlock(info, bytesDecompressed / totalSize.doubleValue);
                                         }
                                     }];
 
@@ -607,6 +637,12 @@ NS_DESIGNATED_INITIALIZER
                                       error:nil];
                         return;
                     }
+
+                    [progress setUserInfoObject:@(++filesExtracted)
+                                         forKey:NSProgressFileCompletedCountKey];
+                    [progress setUserInfoObject:@(fileInfo.count)
+                                         forKey:NSProgressFileTotalCountKey];
+                    progress.completedUnitCount = bytesDecompressed;
                 }
             }
         }
@@ -625,6 +661,13 @@ NS_DESIGNATED_INITIALIZER
 }
 
 - (nullable NSData *)extractData:(UZKFileInfo *)fileInfo
+                           error:(NSError * __autoreleasing*)error
+{
+    return [self extractDataFromFile:fileInfo.filename
+                               error:error];
+}
+
+- (nullable NSData *)extractData:(UZKFileInfo *)fileInfo
                         progress:(void (^)(CGFloat))progress
                            error:(NSError * __autoreleasing*)error
 {
@@ -634,7 +677,18 @@ NS_DESIGNATED_INITIALIZER
 }
 
 - (nullable NSData *)extractDataFromFile:(NSString *)filePath
-                                progress:(void (^)(CGFloat))progress
+                                   error:(NSError * __autoreleasing *)error
+{
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
+    return [self extractDataFromFile:filePath
+                            progress:nil
+                               error:error];
+#pragma clang diagnostic pop
+}
+
+- (nullable NSData *)extractDataFromFile:(NSString *)filePath
+                                progress:(void (^)(CGFloat))progressBlock
                                    error:(NSError * __autoreleasing*)error
 {
     UZKCreateActivity("Extracting Data from File");
@@ -649,16 +703,16 @@ NS_DESIGNATED_INITIALIZER
                                               action:^(NSData *dataChunk, CGFloat percentDecompressed) {
                                                   UZKLogDebug("Appending data chunk of size %lu (%f%% complete)", (unsigned long)dataChunk.length, (double)percentDecompressed);
 
-                                                  if (progress) {
-                                                      progress(percentDecompressed);
+                                                  if (progressBlock) {
+                                                      progressBlock(percentDecompressed);
                                                   }
                                                   
                                                   [result appendData:dataChunk];
                                               }];
     
-    if (progress) {
+    if (progressBlock) {
         UZKLogDebug("Declaring extraction progress as completed");
-        progress(1.0);
+        progressBlock(1.0);
     }
     
     if (success) {
@@ -694,6 +748,8 @@ NS_DESIGNATED_INITIALIZER
         return NO;
     }
     
+    NSProgress *progress = [self beginProgressOperation:fileInfo.count];
+    
     UZKLogInfo("Sorting file info by name/path");
 
     NSArray *sortedFileInfo = [fileInfo sortedArrayUsingDescriptors:@[[NSSortDescriptor sortDescriptorWithKey:@"filename" ascending:YES]]];
@@ -704,15 +760,29 @@ NS_DESIGNATED_INITIALIZER
         BOOL stop = NO;
 
         for (UZKFileInfo *info in sortedFileInfo) {
+            if (progress.isCancelled) {
+                UZKLogInfo("File info iteration was cancelled");
+                break;
+            }
             UZKLogDebug("Performing action on %{public}@", info.filename);
             action(info, &stop);
-            
+            progress.completedUnitCount += 1;
+
             if (stop) {
                 UZKLogInfo("Action dictated an early stop");
+                progress.completedUnitCount = progress.totalUnitCount;
                 break;
             }
         }
     } inMode:UZKFileModeUnzip error:error];
+    
+    if (progress.isCancelled) {
+        NSString *detail = NSLocalizedStringFromTableInBundle(@"User cancelled operation", @"UnzipKit", _resources, @"Detailed error string");
+        UZKLogError("UZKErrorCodeUserCancelled: %{public}@", detail);
+        [self assignError:error code:UZKErrorCodeUserCancelled
+                   detail:detail];
+        return NO;
+    }
     
     return success;
 }
@@ -758,6 +828,8 @@ NS_DESIGNATED_INITIALIZER
 {
     UZKCreateActivity("Extracting Data into Buffer");
     
+    NSProgress *progress = [self beginProgressOperation:0];
+    
     __weak UZKArchive *welf = self;
     const NSUInteger bufferSize = 4096; //Arbitrary
     
@@ -779,6 +851,8 @@ NS_DESIGNATED_INITIALIZER
             return;
         }
         
+        progress.totalUnitCount = info.uncompressedSize;
+
         UZKLogInfo("Opening file");
         if (![welf openFile:innerError]) {
             UZKLogError("Failed to open file %{public}@ in archive", filePath);
@@ -789,6 +863,11 @@ NS_DESIGNATED_INITIALIZER
         
         for (;;)
         {
+            if (progress.isCancelled) {
+                UZKLogInfo("Buffered data read cancelled");
+                return;
+            }
+            
             @autoreleasepool {
                 UZKLogDebug("Reading file data");
                 NSMutableData *data = [NSMutableData dataWithLength:bufferSize];
@@ -816,6 +895,8 @@ NS_DESIGNATED_INITIALIZER
                     UZKLogDebug("Performing action on chunk of data");
                     action([data copy], bytesDecompressed / (CGFloat)info.uncompressedSize);
                 }
+                
+                progress.completedUnitCount = bytesDecompressed;
             }
         }
         
@@ -833,6 +914,14 @@ NS_DESIGNATED_INITIALIZER
             return;
         }
     } inMode:UZKFileModeUnzip error:error];
+    
+    if (progress.isCancelled) {
+        UZKLogError("User cancelled data extraction");
+        NSString *detail = NSLocalizedStringFromTableInBundle(@"User cancelled data read", @"UnzipKit", _resources, @"Detailed error string");
+        [self assignError:error code:UZKErrorCodeUserCancelled
+                   detail:detail];
+        return NO;
+    }
     
     return success;
 }
@@ -894,7 +983,6 @@ NS_DESIGNATED_INITIALIZER
     UZKLogDebug("Decrypting smallest file in archive: %{public}@", smallest.filename);
     
     NSData *smallestData = [self extractData:(UZKFileInfo* _Nonnull)smallest
-                                    progress:nil
                                        error:&error];
     
     if (error || !smallestData) {
@@ -920,7 +1008,6 @@ NS_DESIGNATED_INITIALIZER
          compressionMethod:UZKCompressionMethodDefault
                   password:nil
                  overwrite:YES
-                  progress:nil
                      error:error];
 }
 
@@ -942,6 +1029,20 @@ NS_DESIGNATED_INITIALIZER
 - (BOOL)writeData:(NSData *)data
          filePath:(NSString *)filePath
          fileDate:(NSDate *)fileDate
+            error:(NSError * __autoreleasing*)error
+{
+    return [self writeData:data
+                  filePath:filePath
+                  fileDate:fileDate
+         compressionMethod:UZKCompressionMethodDefault
+                  password:nil
+                 overwrite:YES
+                     error:error];
+}
+
+- (BOOL)writeData:(NSData *)data
+         filePath:(NSString *)filePath
+         fileDate:(NSDate *)fileDate
          progress:(void (^)(CGFloat percentCompressed))progress
             error:(NSError * __autoreleasing*)error
 {
@@ -952,6 +1053,22 @@ NS_DESIGNATED_INITIALIZER
                   password:nil
                  overwrite:YES
                   progress:progress
+                     error:error];
+}
+
+- (BOOL)writeData:(NSData *)data
+         filePath:(NSString *)filePath
+         fileDate:(NSDate *)fileDate
+compressionMethod:(UZKCompressionMethod)method
+         password:(NSString *)password
+            error:(NSError * __autoreleasing*)error
+{
+    return [self writeData:data
+                  filePath:filePath
+                  fileDate:fileDate
+         compressionMethod:UZKCompressionMethodDefault
+                  password:password
+                 overwrite:YES
                      error:error];
 }
 
@@ -979,7 +1096,28 @@ compressionMethod:(UZKCompressionMethod)method
 compressionMethod:(UZKCompressionMethod)method
          password:(NSString *)password
         overwrite:(BOOL)overwrite
-         progress:(void (^)(CGFloat percentCompressed))progress
+            error:(NSError * __autoreleasing*)error
+{
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
+    return [self writeData:data
+                  filePath:filePath
+                  fileDate:fileDate
+         compressionMethod:UZKCompressionMethodDefault
+                  password:password
+                 overwrite:overwrite
+                  progress:nil
+                     error:error];
+#pragma clang diagnostic pop
+}
+
+- (BOOL)writeData:(NSData *)data
+         filePath:(NSString *)filePath
+         fileDate:(NSDate *)fileDate
+compressionMethod:(UZKCompressionMethod)method
+         password:(NSString *)password
+        overwrite:(BOOL)overwrite
+         progress:(void (^)(CGFloat percentCompressed))progressBlock
             error:(NSError * __autoreleasing*)error
 {
     UZKCreateActivity("Writing Data");
@@ -987,16 +1125,19 @@ compressionMethod:(UZKCompressionMethod)method
     UZKLogInfo("Writing data to archive. filePath: %{public}@, fileDate: %{time_t}ld, compressionMethod: %ld, password: %{public}@, "
                "overwrite: %{public}@, progress block specified: %{public}@, error pointer specified: %{public}@",
                filePath, lrint(fileDate.timeIntervalSince1970), (long)method, password != nil ? @"<specified>" : @"(null)", overwrite ? @"YES" : @"NO",
-               progress ? @"YES" : @"NO", error ? @"YES" : @"NO");
+               progressBlock ? @"YES" : @"NO", error ? @"YES" : @"NO");
     
     const NSUInteger bufferSize = 4096; //Arbitrary
     const void *bytes = data.bytes;
     
-    if (progress) {
+    NSProgress *progress = [self beginProgressOperation:data.length];
+    progress.cancellable = NO;
+    
+    if (progressBlock) {
         UZKLogDebug("Calling progress block with zero");
-        progress(0);
+        progressBlock(0);
     }
-
+    
     __weak UZKArchive *welf = self;
     uLong calculatedCRC = crc32(0, data.bytes, (uInt)data.length);
     UZKLogDebug("Calculated CRC: %010lu", calculatedCRC);
@@ -1021,10 +1162,12 @@ compressionMethod:(UZKCompressionMethod)method
                 return err;
             }
             
-            if (progress) {
+            progress.completedUnitCount += size;
+            
+            if (progressBlock) {
                 double percentComplete = i / (double)data.length;
                 UZKLogDebug("Calling progress block at %f%%", percentComplete);
-                progress(percentComplete);
+                progressBlock(percentComplete);
             }
         }
         
@@ -2518,6 +2661,29 @@ compressionMethod:(UZKCompressionMethod)method
     
     UZKLogDebug("Compression method: %lu", file_info.compression_method);
     return file_info.compression_method == 9;
+}
+
+- (NSProgress *)beginProgressOperation:(unsigned long long)totalUnitCount
+{
+    UZKCreateActivity("-beginProgressOperation:");
+    
+    NSProgress *progress;
+    progress = self.progress;
+    self.progress = nil;
+    
+    if (!progress) {
+        progress = [[NSProgress alloc] initWithParent:[NSProgress currentProgress]
+                                             userInfo:nil];
+    }
+    
+    if (totalUnitCount > 0) {
+        progress.totalUnitCount = totalUnitCount;
+    }
+    
+    progress.cancellable = YES;
+    progress.pausable = NO;
+    
+    return progress;
 }