+++ Tests/ExtractDataTests.m
@@ -41,11 +41,6 @@
             
             NSError *error = nil;
             NSData *extractedData = [archive extractDataFromFile:expectedFilename
-                                                        progress:^(CGFloat percentDecompressed) {
-#if DEBUG
-                                                            UZKLogDebug("Extracting, %f%% complete", percentDecompressed);
-#endif
-                                                        }
                                                            error:&error];
             
             XCTAssertNil(error, @"Error in extractData:error:");
@@ -57,11 +52,6 @@
             
             error = nil;
             NSData *dataFromFileInfo = [archive extractData:fileInfos[i]
-                                                   progress:^(CGFloat percentDecompressed) {
-#if DEBUG
-                                                       UZKLogDebug("Extracting from file info, %f%% complete", percentDecompressed);
-#endif
-                                                   }
                                                       error:&error];
             XCTAssertNil(error, @"Error extracting data by file info (%@)", testArchiveName);
             XCTAssertTrue([expectedFileData isEqualToData:dataFromFileInfo], @"Extracted data from file info doesn't match original file (%@)", testArchiveName);
@@ -86,11 +76,6 @@
         
         NSError *error = nil;
         NSData *extractedData = [archive extractDataFromFile:expectedFilename
-                                                    progress:^(CGFloat percentDecompressed) {
-#if DEBUG
-                                                        UZKLogDebug("Extracting, %f%% complete", percentDecompressed);
-#endif
-                                                    }
                                                        error:&error];
         
         XCTAssertNil(error, @"Error in extractData:error:");
@@ -102,11 +87,6 @@
         
         error = nil;
         NSData *dataFromFileInfo = [archive extractData:fileInfos[i]
-                                               progress:^(CGFloat percentDecompressed) {
-#if DEBUG
-                                                   UZKLogDebug("Extracting from file info, %f%% complete", percentDecompressed);
-#endif
-                                               }
                                                   error:&error];
         XCTAssertNil(error, @"Error extracting data by file info");
         XCTAssertTrue([expectedFileData isEqualToData:dataFromFileInfo], @"Extracted data from file info doesn't match original file (%@)", expectedFilename);
@@ -122,11 +102,6 @@
         
         NSError *error = nil;
         NSData *data = [archive extractDataFromFile:@"Test File A.txt"
-                                           progress:^(CGFloat percentDecompressed) {
-#if DEBUG
-                                               UZKLogDebug("Extracting, %f%% complete", percentDecompressed);
-#endif
-                                           }
                                               error:&error];
         
         XCTAssertNotNil(error, @"Extract data without password succeeded");
@@ -141,11 +116,6 @@
     
     NSError *error = nil;
     NSData *data = [archive extractDataFromFile:@"Any file.txt"
-                                       progress:^(CGFloat percentDecompressed) {
-#if DEBUG
-                                           UZKLogDebug("Extracting, %f%% complete", percentDecompressed);
-#endif
-                                       }
                                           error:&error];
     
     XCTAssertNotNil(error, @"Extract data for invalid archive succeeded");