+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.9
 
+* Added support for `NSProgress` and `NSProgressReporting` in all extraction and iteration methods (Issue #32)
 * Added detailed logging using new unified logging framework. See [the readme](README.md) for more details (Issue #47)
 * Upgraded project to Xcode 9 and to the macOS 10.13 and iOS 11 SDKs (Issue #61)
 * Consolidated targets so they're shared between iOS and macOS (Issue #62)